Copy submit element to preserve __currentPage value
As mentioned in #5, the validation is not triggered server side, because of missing `__currentPage ` parameter. The `form.submit();` doesn't send the value of the button.

This change copies the name and value of the button into a hidden input field to preserve the name/value and to get it sent through the request.